🔨 update schema database config

+++ prisma/migrations/20211223225226_initial_migrate/migration.sql
@@ -3,7 +3,7 @@ CREATE TABLE "Product" (
     "id" TEXT NOT NULL,
     "name" TEXT NOT NULL,
     "description" TEXT NOT NULL,
-    "price" INTEGER NOT NULL,
+    "price" DOUBLE PRECISION NOT NULL,
     "category" TEXT NOT NULL,
     "categoryId" TEXT NOT NULL,
     "sku" TEXT NOT NULL,
